Fix error running on windows (#201)
On windows was reported the error:

```
Unable to locate executable file: main_windows.exe. Please verify either the file path exists or the file can be found within a directory specified by the PATH environment variable. Also verify the file has a valid extension for an executable file.
```

This is because windows has a different executioon path due to the
requirement of the `.exe` extension.
In thiis execution path, the executable name wa returned as relative
while in the other ones, iit was returned as relative